whisper_server/requirements.txt
Flask>=3.1.0,<4
openai-whisper>=20231117
